requirements.txt
asgiref==3.9.1
Django==5.2.4
psycopg==3.2.9
python-decouple==3.8
sqlparse==0.5.3
